Refactor Chatbot component to use ReactMarkdown for message rendering and move playAudio function to helper
